requirements.txt
amrlib==0.7.1
beautifulsoup4==4.12.2
fitz==0.0.1.dev2
numpy==1.26.2
openai==0.28
pdf2image==1.16.3
penman==1.3.0
pytesseract==0.3.10
PyYAML==6.0.1
Requests==2.31.0
scikit_learn==1.3.2
sentence_transformers==2.2.2
tiktoken==0.5.2
python-pptx==0.6.23
PyMuPDF==1.23.8
PyMuPDFb==1.23.7
pytesseract==0.3.10
dash-bootstrap-components==1.5.0
dash==2.14.2